Added ServiceImpl, and several methods to Swimlane and SwimlaneStatus
