+++ src/main/java/com/revature/ScrumHub/Service/SwimlaneService.java
@@ -1,27 +1,16 @@
 package com.revature.ScrumHub.Service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import java.util.List;
 
-import com.revature.ScrumHub.Repository.SwimlaneRepository;
 import com.revature.ScrumHub.bean.Swimlane;
 
-@Service
-public class SwimlaneService {
+public interface SwimlaneService {
 
+	public Swimlane getSwimlane (int slId);
 	
-	@Autowired
-	SwimlaneRepository swimRepo;
+	public List<Swimlane> getAllSwimlanes(int slId);
 	
-	public Swimlane getSwimlane (int slId) {
-		System.out.println("Service -getSwimlane");
-		return swimRepo.findOne(slId);
-	}
+	public Swimlane createSwimlane (Swimlane swimlane);
 	
-	public Swimlane createSwimlane (Swimlane swimlane) {
-		System.out.println("Service -create Swimlane");
-		swimRepo.save(swimlane);
-		return swimlane;
-	}
-	
+	public void deleteSwimlane (int slId);
 }
